Create a ContentProvider class
The ContentProvider class provides an API for retrieving content from
an eBook and is a wrapper for the EBook class. The main reason for
creating this class is to add MIME type recognition when loading content.
URL transformation may be added in the future.

However, a method for getting the URL scheme is added
to the EBook class.